Add crackable module
