+++ test/shiftable_test.rb
@@ -1,0 +1,7 @@
+require_relative './test_helper'
+require './lib/shift'
+require 'date'
+
+class ShiftableTest < MiniTest::Test
+  
+end
